pyproject.toml
[project]
name = "textual-pdf-view"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "NSPC911" }]
dependencies = [
  "click>=8.2.1",
  "Pillow>=11.3.0",
  "PyMuPDF>=1.21.64",
  "textual-image>=0.8.3",
  "textual>=5.3.0",
]

[project.scripts]
textual-pdf-view = "textual_pdf_view:main"

[dependency-groups]
dev = [
  "poethepoet>=0.37.0",
  "ruff>=0.12.10",
  "textual-dev>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.poe.tasks]
run = "textual_pdf_view:main"
[tool.ruff.lint]
select = [
  "ASYNC251",
  "ANN",
  "COM819",
  "C400",
  "DOC",
  "D404",
  "E",
  "F",
  "I",
  "N801", "N802", "N805",
  "PLE1142",
  "Q",
  "SIM",
  "TD",
  "W",
]
ignore = ["W505", "E501", "ANN002", "ANN003", "TD002", "TD003"]
preview = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
preview = true

